requirements.txt

Flask==3.0.3
python-dotenv==1.0.1
pydantic==2.8.2
requests==2.32.3
waitress==3.0.0
pytest==8.2.0
